# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Made with ❤ by akatgelar.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "akatgelar@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/dashboard/top-page": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get top page",
                "operationId": "dashboard_top_page",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-device": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get top device",
                "operationId": "dashboard_top_device",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-os": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get top os",
                "operationId": "dashboard_top_os",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-browser": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get top browser",
                "operationId": "dashboard_top_browser",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-country": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get top country",
                "operationId": "dashboard_top_country",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-city": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get top city",
                "operationId": "dashboard_top_city",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/access-daily": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get access daily",
                "operationId": "dashboard_access_daily",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-monthly": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "",
                "description": "Get top monthly",
                "operationId": "dashboard_top_monthly",
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": [],
                                    "metadata": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}